Move prop middleware to src/app/propMiddleware
