+++ pages/about.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Link from 'next/link'
 import Page from '../components/Page'
-import withCommonProps from '../src/withCommonProps'
+import withCommonProps from '../src/app/propMiddleware/withCommonProps'
 
 export default function About ({ indexUrl }) {
   return (
