+++ pages/index.js
@@ -3,7 +3,7 @@ import Link from 'next/link'
 import Page from '../components/Page'
 import Thoughts from '../components/Thoughts'
 import { fetchRecentThoughts } from '../src/app/factory'
-import withCommonProps from '../src/withCommonProps'
+import withCommonProps from '../src/app/propMiddleware/withCommonProps'
 
 export default function Index ({ archiveUrl, thoughts }) {
   return (
