+++ pages/thoughts/archive.js
@@ -3,7 +3,7 @@ import Link from 'next/link'
 import Page from '../../components/Page'
 import Thoughts from '../../components/Thoughts'
 import { fetchAllThoughts } from '../../src/app/factory'
-import withCommonProps from '../../src/withCommonProps'
+import withCommonProps from '../../src/app/propMiddleware/withCommonProps'
 
 export default function ThoughtArchive ({ indexUrl, thoughts }) {
   return (
